Improved subtree handling
Previously, we just built an op_next chain that was virtually the same
as op_sibling of the JITOP children. That's not good enough. A child has
to point at the left-most OP of the next child.

Currently, we emit a CONSTOP as child if it's the very left-most OP in
the entire subtree. This is to capture op_next pointers from outside the
subtree. If we null the CONSTOP, those pointers go away for some reason.