Temporary -- slow -- fix for modification of RO value
This relates to dATARGET (or its unrolled equivalent) occasionally
choosing either some PAD or the top element on the stack as output for a
given operation. If that target ends up being RO, then *bang*. For now,
the workaround is to push a new and mortalized SV on the stack, but that
SUCKS badly in terms of performance. For now: working trumps working
fast.