Expose && and || to Perl; learn to skip OP_NULL
Exposes the && and || implementations to Perl. Only recognizes them as
part of a subtree, not as a root OP since they can appear in scary
places like loops.
These OPs can occasionally cause an OP_NULL to be included in the tree,
so for now, we're skipping all OP_NULLs within out subtree by pretending
they're UNOPs. That's probably very fragile.