Added a new Term type that is jiittable and jiitted to nothing.
Attributes on lexical variables are evaluated at runtime, so
the perl parser actually emits an optree for the method call registering
the attributes:

    my $x : Int;

expands to something equivalent to:

    my $x;
    attributes::->import(__PACKAGE__, \$x, 'Int');

and the call happens at runtime every time execution reaches
the lexical declaration.

This is BAD, especially for performance.

The next commit will add code that recognizes the optree generated by attribute
expansion, parses the attributes and returns a NullOptree term that will be
replaced with a stub op by the JIT (or simply jitted to nothing when we manage to
jit all the code around it).

It might make sense to attach the attributes to the NullOptree node, in the
spirit of providing the caller with more rope to han... more information about
the syntax tree.