WIP: Keyword plugin for "typed Integer $foo"
This isn't working at all.

The actual lexicalization of the keyword plugin as such works alright,
as does the parsing all the way up to the list of variables.

While Perl does expose a "parse this as a list expression" parser API
that can build an OP tree for us, it's obviously not the same as
parsing the list following a "my". So the OP tree returned includes a
bunch of global variable accesses where normally, it would have a
bunch of PADSV OPs. Not entirely surprising, I know.

The way out is to actually parse that expression and build the OP tree
ourselves. That's including a big chunk of perl-version-implementation
specific logic into our project. Potentially, a slightly more limited
form of the full my semantics would be doable?

To wit, perly.y says (skipped code):

  myattrterm:	MY myterm myattrlist
          |	MY myterm
          ;

  myterm	:	'(' expr ')'
          |	'(' ')'
          |	scalar	%prec '('
          |	hsh 	%prec '('
          |	ary 	%prec '('
          ;

  scalar	:	'$' indirob
  ary	:	'@' indirob
  hsh	:	'%' indirob

  indirob	:	WORD
          |	scalar %prec PREC_LOW
          |	block
          |	PRIVATEREF
          ;

If we drop %prec, and myattrlist, and the fancier variants of indirob,
then this seems quite tractable (if highly tedious) even without a
parser generator or explicit state machine.

Thoughts?