KIDS_OPTIONAL flag may mean function-specific defaults
Now, if the PJ_ASTf_KIDS_OPTIONAL flag is set, kids aren't just optional
in the AST, but the exact desired behaviour is going to depend greatly
on the actual operation. "abs()" is going to use $_. "rand()" is going
to use 1. That kind of logic will have to be implemented during code
generation for now, but could conceivably be done during AST generation
as well. But that would be more work now, for unclear benefit (and also
pose the problem of how to represent $_ while we can't really handle
global variables yet. Also $_ may be lexical? Ugh).

Review desired.