Add a StatementSequence class.
It's just a list of consecutive statements; at some point it will
be replaced by a proper AST for all constructs that contain one or
more blocks as part of their definition.

In the meantime, knowing that one or more statement are consecutive
allows implementing analysis code for variable access that is applicable
to all sequence structures (hopefully with minimal changes) and allows
implementing some more aggressive (and hopefully more effective)
optimizations.