Add "List" term type
In a nutshell, this is a low-on-semantics structure for grouping things.
An array assignment has two lists: Of lvalues and rvalues. A list slice
does, too: Of values and indices. Instead of creating an Op type that has
two or more lists of kid-ops, this is just a container to hold more children.

It is not clear that this is the best idea since sliced bread, but it should
be easy enough to back out of.