Add more types used in Emit.pm
ANY can only be used in the requested type for an expression, SV
and GV can be used both as requested type and returned type and map
to the native Perl values.

There are no AV and HV for now (it's better to wait for an actual use
case, it might be that Array/Hash types need some refactroing anyway).