AST: aelem and aelemfast
What makes this interesting is that aelemfast can refer to lexicals and globals.
On top of that, it means that now AST::Lexical and AST::Global terms can now
point to OPs of different types. For accessing the respective lexical or global,
the different OP is expected to behave the same as the OP_PADXV or OP_GV.