Naive attempt to add scope/enter/leave to AST
The thing this runs afoul of is the StatementSequence and Statement
generating logic. StatementSequence's are assumed to be at the very top
level and at the very top level only, so the op tree walker keeps one
around to append each statement to. This could be converted into more
of a stack or in a number of other ways. Not sure what the best solution
is at this point. :(