Add emitter skeleton using Algorithm::Burs::Cpp as the dispatcher
Does not JIT any operation yet, but it contains all the support code required
to handle non-jittable optrees (containing opaque value, non-handled opcodes or
an op we don't want to JIT in isolation, e.g. padsv).

Emitter code for binary ops, lexicals, constants in pj_emitter.cpp is currently
unused, but it's left there to reduce the size of the diff.